Now crops to multiples of 4 pixels when autodetecting crop region.
Script removes temp files when done.

Changed the default temp dir to the presend working directory.

Fixed the help message.  Apparently it was due to a bug in Getopt::Declare
where you couldnt end a pvtype regex with a '+'.  This got even more fed up
with Getopt::Declare.

Changed how the nice-ing works.  Now the default is always 10 and the switch
requires an argument.

Now we handle failed system calls and abort the script instead of blindly
trudging forward.

The script now checks audio size and length before trying to calculate a bogus
bitrate.