Major version bump.
I made the necessary changes to make the system calls avoid using a shell.
Many things got changed to support this.  Some variables where turned to
arrays, escaping of filenames was removed, my_system function was re-written,
and all the system calls were re-written for this reason.  Also, backticks were
replaced with special open calls to support the new unescaped filenames, and
the script now forks to run oggenc.  Now everything should be much happier (tm)
and we should be able to handle any crazy special characters in filenames that
get throw at us.

More information is printed out after the audio stream is encoded.

Ripping to a target filesize should now be more accurate (especially on small
target sizes) as we now sleep for a second to allow oggenc to finish encoding
whats in the pipe before checking the audio length and size.