Now we fork, exec, and wait for our child instead of just running commands with system.  This allows use to shut up stderr by closing that stream.  Unfortunat- ly, it looks like oggenc gives it progress info on stderr so I may have to look into handling that differently.
I also want to think about how I'm handling forking for the oggenc call so that
the pretend output is guaranteed to be correct.  

Finally, we should do some sort of error checking on the exec calls that we
make but I'm not really sure how to do that now that they are in child
processes.  I would have to know more about IPC for that.