Fixed (hopefully?) dvd ripping filesize problem by adding the '-ofps' flag to rips from dvds.
Switched the mencoder lines to use pcm audio instead of copy.  This makes
ripping things with audio that mencoder doesn't grok work at the cost of much
larger temp files.

Fixed a bug where temp files were kept when the final file size was right and
not kept when it was wrong.  Whoops ('>' != '<').