Several changes to dvdrip:
If '--' is given on the command line, all following options are passed directly 
to mplayer/mencoder.  For example, "dvdrip ... -- -chapter 3-4" will cause only 
chapters 3 and 4 to be extracted from a DVD.


If -k is given, don't print "Final file size off by more than 2%, keeping temp 
files".  However, continue to keep the files.


Default audio track to 128 (previously left to mplayer).  This was necessitated 
by the next change.


Allow multiple audio tracks (e.g., "-aid 128,129").  This requires generating a 
different ogg file for each track ("audio-$pid-$aid.ogg").  The default audio 
id makes this simpler since it makes the case for several audio ids the same as 
the case with one audio id (for loop over either one or many elements).  Also, 
in instances in which the audio track is not relevant but is specified anyway 
(e.g., bitrate), we use the first one since we know there is at least one.

Note: the first aid given is the default aid.